requirements.txt
Flask==3.1.1
openai>=1.30.0
python-dotenv==1.1.1
python-docx==1.2.0
PyPDF2==3.0.1
Pillow==11.2.1
upstash-redis==1.4.0
lxml==5.4.0
Markdown==3.8.2
MarkupSafe==3.0.2
Jinja2==3.1.6
Werkzeug==3.1.3
requests==2.32.4
